requirements/shared.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
PyYAML>=6.0.0
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
pandas>=2.1.0
websocket-client>=1.6.0